+++ src/super/i-input/CHANGELOG.md
@@ -14,6 +14,11 @@ Changelog
 #### :rocket: New Feature
 
 * Added support of interpolation of a data provider response
+* Added `isFocused`
+
+#### :nail_care: Polish
+
+* Improved documentation
 
 ## v3.0.0-rc.49 (2020-08-03)